+++ packages/components/organisms/f-checkout/src/constants.js
@@ -22,6 +22,7 @@ export const VALIDATIONS = {
 
 export const VUEX_CHECKOUT_MODULE = 'fCheckoutModule';
 export const VUEX_CHECKOUT_ANALYTICS_MODULE = 'fCheckoutAnalyticsModule';
+export const VUEX_CHECKOUT_EXPERIMENTATION_MODULE = 'fCheckoutExperimentationModule';
 
 export const ERROR_CODE_FULFILMENT_TIME_INVALID = 'FULFILMENT_TIME_INVALID';
 export const ERROR_CODE_FULFILMENT_TIME_UNAVAILABLE = 'FULFILMENT_TIME_UNAVAILABLE';
@@ -37,3 +38,5 @@ export const CHEKOUT_ERROR_FORM_TYPE = {
     default: 'pageLoad',
     accessForbidden: 'accessForbiddenError'
 };
+
+export const HEADER_LOW_VALUE_ORDER_EXPERIMENT = 'x-je-exp-low_value_order_threshold_web';