+++ packages/components/organisms/f-checkout/src/store/checkoutAnalytics.module.js
@@ -1,12 +1,7 @@
-import {
-    mapAnalyticsName, mapAnalyticsNames, getAnalyticsErrorCodeByApiErrorCode
-} from '../services/mapper';
-import { VUEX_CHECKOUT_MODULE } from '../constants';
-
-import {
-    UPDATE_AUTOFILL,
-    UPDATE_CHANGED_FIELD
-} from './mutation-types';
+import { mapAnalyticsName, mapAnalyticsNames, getAnalyticsErrorCodeByApiErrorCode } from '../services/mapper';
+import experimentService from '../services/experimentService';
+import { VUEX_CHECKOUT_MODULE, HEADER_LOW_VALUE_ORDER_EXPERIMENT } from '../constants';
+import { UPDATE_AUTOFILL, UPDATE_CHANGED_FIELD } from './mutation-types';
 
 export default {
     namespaced: true,
@@ -111,7 +106,7 @@ export default {
         },
 
         /**
-         *Dispatches `trackFormInteraction` with each error in `state.errors`.
+         * Dispatches `trackFormInteraction` with each error in `state.errors`.
          */
         trackFormErrors ({ rootState, dispatch }) {
             const trackedErrors = [];
@@ -125,6 +120,17 @@ export default {
                     dispatch('trackFormInteraction', { action: 'error', error: mappedError });
                 }
             });
+        },
+
+        /**
+         * Fetches the variant of the Low Value Order experiment from the headers and pushes an analytics event.
+         */
+        trackLowValueOrderExperiment: experimentHeaders => {
+            const lowValueOrderExperimentVariant = experimentHeaders[HEADER_LOW_VALUE_ORDER_EXPERIMENT];
+            const event = experimentService.getLowValueOrderExperimentTracking(lowValueOrderExperimentVariant);
+            if (event) {
+                window.dataLayer.push(event);
+            }
         }
     },