+++ packages/components/organisms/f-checkout/src/store/tests/checkout.module.test.js
@@ -364,11 +364,16 @@ describe('CheckoutModule', () => {
         let commit;
         let dispatch;
         let payload;
-
+        let rootGetters;
+        let context;
         beforeEach(() => {
             commit = jest.fn();
             dispatch = jest.fn();
             state = defaultState;
+            rootGetters = {};
+            context = {
+                commit, state, dispatch, rootGetters
+            };
             payload = {
                 url: 'http://localhost/account/checkout',
                 getCustomerUrl: 'http://localhost/customer',
@@ -400,7 +405,7 @@ describe('CheckoutModule', () => {
 
             it(`should get the checkout details from the backend and call ${UPDATE_STATE} mutation.`, async () => {
                 // Act
-                await getCheckout({ commit, state, dispatch }, payload);
+                await getCheckout(context, payload);
 
                 // Assert
                 expect(axios.get).toHaveBeenCalledWith(payload.url, config);
@@ -412,7 +417,7 @@ describe('CheckoutModule', () => {
                 const expectedAsapValue = checkoutDeliveryCopy.fulfilment.time.asap;
 
                 // Act
-                await getCheckout({ commit, state, dispatch }, payload);
+                await getCheckout(context, payload);
 
                 // Assert
                 expect(commit).toHaveBeenCalledWith(UPDATE_HAS_ASAP_SELECTED, expectedAsapValue);
@@ -420,7 +425,7 @@ describe('CheckoutModule', () => {
 
             it(`should call '${VUEX_CHECKOUT_ANALYTICS_MODULE}/updateAutofill' mutation with an array of updated field names.`, async () => {
                 // Act
-                await getCheckout({ commit, state, dispatch }, payload);
+                await getCheckout(context, payload);
 
                 // Assert
                 expect(dispatch).toHaveBeenCalledWith(`${VUEX_CHECKOUT_ANALYTICS_MODULE}/updateAutofill`, state, { root: true });
@@ -432,7 +437,7 @@ describe('CheckoutModule', () => {
                     checkoutDeliveryCopy.customer = null;
 
                     // Act
-                    await getCheckout({ commit, state, dispatch }, payload);
+                    await getCheckout(context, payload);
 
                     // Assert
                     expect(checkoutDeliveryCopy.customer).toBe(null);
@@ -450,7 +455,7 @@ describe('CheckoutModule', () => {
                     };
 
                     // Act
-                    await getCheckout({ commit, state, dispatch }, payload);
+                    await getCheckout(context, payload);
 
                     // Assert
                     expect(checkoutDeliveryCopy.customer.phoneNumber).toBe(expectedPhoneNumber);
@@ -466,7 +471,7 @@ describe('CheckoutModule', () => {
                     };
 
                     // Act
-                    await getCheckout({ commit, state, dispatch }, payload);
+                    await getCheckout(context, payload);
 
                     // Assert
                     expect(checkoutDeliveryCopy.customer.firstName).toBe(expectedFirstName);
@@ -492,7 +497,7 @@ describe('CheckoutModule', () => {
                     const expectedPhoneNumber = '9876543210'; // AuthToken Mobile No.
 
                     // Act
-                    await getCheckout({ commit, state, dispatch }, payload);
+                    await getCheckout(context, payload);
 
                     // Assert
                     expect(checkoutDeliveryCopy.customer.phoneNumber).toBe(expectedPhoneNumber);
@@ -505,7 +510,7 @@ describe('CheckoutModule', () => {
                     const expectedPhoneNumber = '0123456789'; // AuthToken Phone No.
 
                     // Act
-                    await getCheckout({ commit, state, dispatch }, payload);
+                    await getCheckout(context, payload);
 
                     // Assert
                     expect(checkoutDeliveryCopy.customer.phoneNumber).toBe(expectedPhoneNumber);
@@ -537,7 +542,7 @@ describe('CheckoutModule', () => {
                     };
 
                     // Act
-                    await getCheckout({ commit, state, dispatch }, payload);
+                    await getCheckout(context, payload);
 
                     // Assert
                     expect(checkoutDeliveryCopy.customer.firstName).toBe(expectedCustomerDetails.firstName);
@@ -558,7 +563,7 @@ describe('CheckoutModule', () => {
                     };
 
                     // Act
-                    await getCheckout({ commit, state, dispatch }, payload);
+                    await getCheckout(context, payload);
 
                     // Assert
                     expect(checkoutDeliveryCopy.customer.firstName).toBe(expectedCustomerDetails.firstName);
@@ -582,7 +587,7 @@ describe('CheckoutModule', () => {
                 axios.get = jest.fn(() => Promise.resolve({ data: basketDelivery }));
 
                 // Act
-                await getBasket({ commit, state, dispatch }, payload);
+                await getBasket(context, payload);
 
                 // Assert
                 expect(axios.get).toHaveBeenCalledWith(payload.url, config);
@@ -616,7 +621,7 @@ describe('CheckoutModule', () => {
                 const [expectedAddress] = customerAddresses.Addresses;
 
                 // Act
-                await getAddress({ commit, state, dispatch }, payload);
+                await getAddress(context, payload);
 
                 // Assert
                 expect(axios.get).toHaveBeenCalledWith(payload.url, config);
@@ -630,7 +635,7 @@ describe('CheckoutModule', () => {
 
             it(`should call '${VUEX_CHECKOUT_ANALYTICS_MODULE}/updateAutofill' mutation with an array of updated field names.`, async () => {
                 // Act
-                await getAddress({ commit, state, dispatch }, payload);
+                await getAddress(context, payload);
 
                 // Assert
                 expect(dispatch).toHaveBeenCalledWith(`${VUEX_CHECKOUT_ANALYTICS_MODULE}/updateAutofill`, state, { root: true });
@@ -737,7 +742,7 @@ describe('CheckoutModule', () => {
                 }));
 
                 // Act
-                await placeOrder({ commit, state }, payload);
+                await placeOrder(context, payload);
 
                 // Assert
                 expect(axios.post).toHaveBeenCalledWith(payload.url, payload.data, config);
@@ -759,7 +764,7 @@ describe('CheckoutModule', () => {
 
                         try {
                             // Act
-                            await placeOrder({ commit, state }, payload);
+                            await placeOrder(context, payload);
                         } catch {
                             // Assert
                             expect(commit).toHaveBeenCalledWith(UPDATE_ERRORS, [message]);
@@ -780,7 +785,7 @@ describe('CheckoutModule', () => {
 
                         try {
                             // Act
-                            await placeOrder({ commit, state }, payload);
+                            await placeOrder(context, payload);
                         } catch {
                             // Assert
                             expect(commit).toHaveBeenCalledWith(UPDATE_ERRORS, []);
@@ -801,7 +806,7 @@ describe('CheckoutModule', () => {
 
                         try {
                             // Act
-                            await placeOrder({ commit, state, dispatch }, payload);
+                            await placeOrder(context, payload);
                         } catch {
                             // Assert
                             expect(dispatch).toHaveBeenCalledWith('updateMessage', message);
@@ -817,7 +822,7 @@ describe('CheckoutModule', () => {
                     axios.post = jest.fn(() => Promise.reject(new Error(errorMessage)));
 
                     // Act
-                    const result = await expect(placeOrder({ commit, state }, payload));
+                    const result = await expect(placeOrder(context, payload));
 
                     // Assert
                     result.rejects.toThrow(errorMessage);
@@ -858,7 +863,7 @@ describe('CheckoutModule', () => {
 
             it('should post the checkout details to the backend.', async () => {
                 // Act
-                await updateCheckout({ commit, state, dispatch }, payload);
+                await updateCheckout(context, payload);
 
                 // Assert
                 expect(axios.patch).toHaveBeenCalledWith(payload.url, payload.data, config);
@@ -866,7 +871,7 @@ describe('CheckoutModule', () => {
 
             it('should convert an unsupported error into a default error.', async () => {
                 // Act
-                await updateCheckout({ commit, state, dispatch }, payload);
+                await updateCheckout(context, payload);
 
                 // Assert
                 expect(commit).toHaveBeenCalledWith(UPDATE_ERRORS, [{ code: DEFAULT_CHECKOUT_ISSUE, shouldShowInDialog: true }]);
@@ -885,7 +890,7 @@ describe('CheckoutModule', () => {
 
                 it('should call `updateErrors` with issue.', async () => {
                     // Act
-                    await updateCheckout({ commit, state, dispatch }, payload);
+                    await updateCheckout(context, payload);
 
                     // Assert
                     expect(commit).toHaveBeenCalledWith(UPDATE_ERRORS, [issue]);
@@ -893,7 +898,7 @@ describe('CheckoutModule', () => {
 
                 it('should call `updateMessage` with first issue.', async () => {
                     // Act
-                    await updateCheckout({ commit, state, dispatch }, payload);
+                    await updateCheckout(context, payload);
 
                     // Assert
                     expect(dispatch).toHaveBeenCalledWith('updateMessage', issue);
@@ -933,7 +938,7 @@ describe('CheckoutModule', () => {
 
             it('should post the create guest user request to the backend.', async () => {
                 // Act
-                await createGuestUser({ commit, state }, payload);
+                await createGuestUser(context, payload);
 
                 // Assert
                 expect(axios.post).toHaveBeenCalledWith(payload.url, payload.data, config);
@@ -941,7 +946,7 @@ describe('CheckoutModule', () => {
 
             it(`should call ${UPDATE_AUTH_GUEST} mutation.`, async () => {
                 // Act
-                await createGuestUser({ commit, state }, payload);
+                await createGuestUser(context, payload);
 
                 // Assert
                 expect(commit).toHaveBeenCalledWith(UPDATE_AUTH_GUEST, authToken);
@@ -964,7 +969,7 @@ describe('CheckoutModule', () => {
 
             it(`should get the checkout available fulfilment details from the backend and call ${UPDATE_AVAILABLE_FULFILMENT_TIMES} mutation.`, async () => {
                 // Act
-                await getAvailableFulfilment({ commit }, payload);
+                await getAvailableFulfilment(context, payload);
 
                 // Assert
                 expect(axios.get).toHaveBeenCalledWith(payload.url, config);
@@ -975,7 +980,7 @@ describe('CheckoutModule', () => {
         describe('updateFulfilmentTime ::', () => {
             it(`should call ${UPDATE_FULFILMENT_TIME} mutation.`, () => {
                 // Act
-                updateFulfilmentTime({ commit, state }, time);
+                updateFulfilmentTime(context, time);
 
                 // Assert
                 expect(commit).toHaveBeenCalledWith(UPDATE_FULFILMENT_TIME, time);
@@ -1010,7 +1015,7 @@ describe('CheckoutModule', () => {
 
                 it(`should get the geo location details from the backend and call ${UPDATE_GEO_LOCATION} mutation.`, async () => {
                     // Act
-                    await getGeoLocation({ commit, state }, payload);
+                    await getGeoLocation(context, payload);
 
                     // Assert
                     expect(axios.post).toHaveBeenCalledWith(payload.url, locationData, config);
@@ -1025,7 +1030,7 @@ describe('CheckoutModule', () => {
 
                 it('should not make api call and should not call mutation.', async () => {
                     // Act
-                    await getGeoLocation({ commit, state }, payload);
+                    await getGeoLocation(context, payload);
 
                     // Assert
                     expect(axios.post).not.toHaveBeenCalled();
@@ -1072,7 +1077,7 @@ describe('CheckoutModule', () => {
                     };
 
                     // Act
-                    await getGeoLocation({ commit, state: newState }, payload);
+                    await getGeoLocation({ ...context, state: newState }, payload);
 
                     // Assert
                     expect(axios.post).not.toHaveBeenCalled();
@@ -1092,7 +1097,7 @@ describe('CheckoutModule', () => {
                     };
 
                     // Act
-                    await getGeoLocation({ commit, state: newState }, payload);
+                    await getGeoLocation({ ...context, state: newState }, payload);
 
                     // Assert
                     expect(commit).toHaveBeenCalledWith(UPDATE_GEO_LOCATION, geoLocationDetails.geometry.coordinates);
@@ -1105,7 +1110,7 @@ describe('CheckoutModule', () => {
                         jest.spyOn(addressService, 'doesAddressInStorageAndFormMatch').mockImplementation(() => false);
 
                         // Act
-                        await getGeoLocation({ commit, state }, payload);
+                        await getGeoLocation(context, payload);
 
                         // Assert
                         expect(setItemSpy).toHaveBeenCalledWith(
@@ -1124,7 +1129,7 @@ describe('CheckoutModule', () => {
             [updateMessage, UPDATE_MESSAGE, message]
         ])('%s should call %s mutation with passed value', (action, mutation, value) => {
             // Act
-            action({ commit, dispatch }, value);
+            action(context, value);
 
             // Assert
             expect(commit).toHaveBeenCalledWith(mutation, value);
@@ -1135,7 +1140,7 @@ describe('CheckoutModule', () => {
             [updateCustomerDetails, customerDetails]
         ])(`%s should dispatch '${VUEX_CHECKOUT_ANALYTICS_MODULE}/updateChangedFields' action with first key of passed value`, (action, value) => {
             // Act
-            action({ commit, dispatch }, value);
+            action(context, value);
 
             const [field] = Object.keys(value);
 
@@ -1160,7 +1165,7 @@ describe('CheckoutModule', () => {
                         jest.spyOn(window.sessionStorage, 'getItem').mockReturnValue(userNote);
 
                         // Act
-                        getUserNote({ dispatch, state });
+                        getUserNote(context);
 
                         // Assert
                         expect(dispatch).toHaveBeenCalledWith('updateUserNote', userNote);
@@ -1173,7 +1178,7 @@ describe('CheckoutModule', () => {
                         jest.spyOn(window.sessionStorage, 'getItem').mockReturnValue(undefined);
 
                         // Act
-                        getUserNote({ dispatch, state });
+                        getUserNote(context);
 
                         // Assert
                         expect(dispatch).not.toHaveBeenCalled();
@@ -1193,7 +1198,7 @@ describe('CheckoutModule', () => {
 
                 it('should not call dispatch', () => {
                     // Act
-                    getUserNote({ dispatch, state });
+                    getUserNote(context);
 
                     // Assert
                     expect(dispatch).not.toHaveBeenCalled();
@@ -1219,7 +1224,7 @@ describe('CheckoutModule', () => {
                 const key = `userNote-${testBasketId}`;
 
                 // Act
-                saveUserNote({ state });
+                saveUserNote(context);
 
                 // Assert
                 expect(spy).toHaveBeenCalledWith(key, state.userNote);